refactor: prefix every unified reports component with Unified
The navigation components kept their pre-migration Reports* names, which read
as generic report chrome inside a domain where everything else is
UnifiedReport*. They become UnifiedReportsMegaMenu,
UnifiedReportsMobileSelectionDrawer, and UnifiedReportsNavigationSidebar
(plural, since they navigate across reports, unlike the singular
UnifiedReport* components that render one). The drawer's stylesheet and
Layer__ class follow the component.

Helpers land where their consumers are: the column and sub-row builders were
all table-specific with UnifiedReportTable as the only caller, so they move
into UnifiedReportTable/utils.tsx, which leaves the navigation config
builders as the domain's single root utils.ts. The report combo box option
class moves beside the mobile selection drawer that constructs it, matching
how the other ComboBoxOption classes sit with their components.
