+++ src/components/features/unifiedReports/UnifiedReportsMobileSelectionDrawer/unifiedReportsMobileSelectionDrawer.scss
@@ -1,4 +1,4 @@
-.Layer__ReportsMobileSelectionDrawer {
+.Layer__UnifiedReportsMobileSelectionDrawer {
   inline-size: 100%;
   max-inline-size: 16rem;
 }
