+++ src/components/features/unifiedReports/UnifiedReportsNavigationSidebar/UnifiedReportsNavigationSidebar.tsx
@@ -7,12 +7,12 @@ import { Span } from '@ui/Typography/Text'
 import { TreeNavigation } from '@blocks/NestedNavigation/TreeNavigation/TreeNavigation'
 import { TreeNavigationSkeleton } from '@blocks/NestedNavigation/TreeNavigation/TreeNavigationSkeleton'
 import { ConditionalBlock } from '@components/utility/ConditionalBlock'
+import { UnifiedReportsNavigationSidebarError } from '@features/unifiedReports/UnifiedReportsNavigationSidebar/UnifiedReportsNavigationSidebarError'
 import {
   buildRecursiveReportsGroupConfig,
   buildReportsLeafConfig,
   isReportGroup,
-} from '@features/unifiedReports/navigationUtils'
-import { ReportsNavigationSidebarError } from '@features/unifiedReports/ReportsNavigationSidebar/ReportsNavigationSidebarError'
+} from '@features/unifiedReports/utils'
 
 const groupConfig = buildRecursiveReportsGroupConfig(
   group => <Span ellipsis weight='bold'>{group.displayName}</Span>,
@@ -22,7 +22,7 @@ const renderLeafLabel = (leaf: { displayName: string }) => (
   <Span ellipsis>{leaf.displayName}</Span>
 )
 
-export function ReportsNavigationSidebar() {
+export function UnifiedReportsNavigationSidebar() {
   const { t } = useTranslation()
   const { data, isLoading, isError } = useGetReportConfig()
   const { baseReport, setBaseReport } = useBaseUnifiedReport()
@@ -39,7 +39,7 @@ export function ReportsNavigationSidebar() {
       Loading={<TreeNavigationSkeleton />}
       Inactive={null}
       isError={isError}
-      Error={<ReportsNavigationSidebarError />}
+      Error={<UnifiedReportsNavigationSidebarError />}
     >
       {({ data }) => (
         <TreeNavigation
